+++ transform/lib/passes/replacer.js
@@ -33,9 +33,7 @@ export class ThrowReplacer extends Visitor {
         super.visitThrowStatement(node, ref);
         const value = node.value;
         const newThrow = Node.createIfStatement(Node.createCallExpression(Node.createIdentifierExpression("isDefined", node.range), null, [Node.createPropertyAccessExpression(node.value, Node.createIdentifierExpression("rethrow", node.range), node.range)], node.range), Node.createExpressionStatement(Node.createCallExpression(Node.createPropertyAccessExpression(node.value, Node.createIdentifierExpression("rethrow", node.range), node.range), null, [], node.range)), Node.createThrowStatement(node.value, node.range), node.range);
-        debugger;
         replaceRef(node, [newThrow], ref);
-        console.log(toString(newThrow));
     }
     static replace(sources) {
         const replacer = new ThrowReplacer();