+++ transform/src/passes/source.ts
@@ -480,7 +480,6 @@ export class SourceLinker extends Visitor {
       const entrySourceRef = Globals.sources.get(entrySource.internalPath)!;
       entrySourceRef.linker.gather();
       entrySourceRef.linker.link(true);
-      debugger;
     }
 
     for (const entrySource of entrySources) {